Rename core app to utils and update imports
Refactor utils/perms.py to accounts
