+++ renthub/urls.py
@@ -17,6 +17,5 @@ Including another URLconf
 from django.urls import path, include
 
 urlpatterns = [
-    path('', include('core.urls')),
     path('', include('admin_site.urls'))
 ]
